feat: Add arguments for deleting frames
Due to the way the id3 crate is implemented, in order to delete a
concrete COMM[desc, lang] or WXXX[desc], etc., it's necessary to first
remove all COMM/WXXX/... frames, then iterate through the removed ones,
and add back the ones that are not relevant.
This is a bit annoying, and it does alter the order of untouched frames.
But it seems to work correctly nevertheless.