feat(cli): Add convert and purge options to --help
Not implemented yet, it's a draft. Inspired by the options offered by
mid3v2, but improved. I think this covers most use cases, and it solves
the problem of operating on files with no pre-existing tags (because now
there's a way to specify which tag version to create, and the default is
also defined as ID3v2.4).