fix: Don't write tag after purge
This made purge just not work, because it would always write the tag
back afterwards.

In the process I also had to enclose the action loop in an if condition
to prevent duplicate prints of 'No tag found'. This is a big ugly, and
the indentation level is becoming crazy. Should probably be refactored.

Issue: #2