test: Fix TXXX and COMM samples encoding
The samples were created with mid3v2, but as it turns out, mid3v2 has a
quirk that causes it to terminate TXXX and COMM frames with a NULL byte.
To get rid of these extra bytes, I reencoded the samples with eyeD3.

This is not well-defined by the standard and the id3 crate currently is
not able to parse the extra NULL byte "correctly" (this is not
necessarily an error according to the standard, but other popular ID3
tools omit the trailing NULL byte).
See https://github.com/polyfloyd/rust-id3/issues/135 for more info.

Issue: #1