fix(cli): Check frame ID lengths on cmdline
Without this, rsid3 interprets something invalid like --XD as a getter
and attempts to create an XD text frame, triggering a panic from within
the id3 crate.

In ID3v2.3 and ID3v2.4, all frames IDs are guaranteed a length of 4.
This will get changed as support for ID3v2.2 gets added later.